feat(front): u#3172 t#3448 wks detail refactor
