+++ javascript/apps/taiga/src/app/modules/auth/feature-login/components/login/login.component.html
@@ -37,8 +37,7 @@ Copyright (c) 2023-present Kaleidos INC
             [state]="{ usernameInput: loginForm.get('username')?.value }"
             class="forgot-password"
             appearance="primary"
-            tuiLink
-            type="button">
+            tuiLink>
             {{ t('login.forgot_password') }}
           </a>
         </ng-container>
