+++ javascript/apps/taiga/src/app/modules/feature-new-project/components/init-step/init-step.component.html
@@ -10,7 +10,6 @@ Copyright (c) 2023-present Kaleidos INC
   <div class="new-step-top-actions">
     <a
       tuiButton
-      type="button"
       appearance="tertiary"
       [routerLink]="getPreviousUrl"
       icon="arrow-left">
