+++ javascript/apps/taiga/src/app/modules/feature-new-project/components/template-step/template-step.component.html
@@ -128,7 +128,7 @@ Copyright (c) 2023-present Kaleidos INC
           tuiButton
           type="submit"
           appearance="primary">
-          {{ t('new_project.second_step.create_project') }}
+          {{ t('common_project.create_project') }}
         </button>
       </div>
     </form>
