+++ javascript/apps/taiga/src/app/modules/project/feature-navigation/components/project-navigation-menu/project-navigation-menu.component.html
@@ -49,7 +49,8 @@ Copyright (c) 2023-present Kaleidos INC
             [routerLink]="[
               '/workspace',
               project.workspace.id,
-              project.workspace.slug
+              project.workspace.slug,
+              'projects'
             ]">
             {{ project.workspace.name }}
           </a>
