+++ javascript/apps/taiga/src/app/modules/workspace/feature-detail/components/workspace-detail-people/workspace-detail-people.component.css
@@ -1,0 +1,10 @@
+/*
+This Source Code Form is subject to the terms of the Mozilla Public
+License, v. 2.0. If a copy of the MPL was not distributed with this
+file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+Copyright (c) 2023-present Kaleidos INC
+*/
+
+@import "tools/typography.css";
+@import "../../../styles/workspace.shared";
