+++ javascript/apps/taiga/src/app/modules/workspace/styles/workspace.shared.css
@@ -76,3 +76,17 @@ Copyright (c) 2023-present Kaleidos INC
     padding-block-end: var(--spacing-16);
   }
 }
+
+.wks-main-content-top {
+  align-items: center;
+  display: flex;
+  justify-content: space-between;
+  padding-block: var(--spacing-8);
+  padding-inline: var(--spacing-20) var(--spacing-8);
+}
+
+.wks-main-content-title {
+  @mixin font-heading-4;
+
+  color: var(--color-gray100);
+}
