+++ javascript/apps/taiga/src/app/shared/navigation/navigation.component.html
@@ -14,8 +14,7 @@ Copyright (c) 2023-present Kaleidos INC
         [routerLink]="['/']"
         class="home-button"
         tuiIconButton
-        icon="taiga-logo"
-        type="button">
+        icon="taiga-logo">
       </a>
     </div>
     <div class="accesibility-help">
@@ -54,8 +53,7 @@ Copyright (c) 2023-present Kaleidos INC
         [routerLink]="['/']"
         class="home-button"
         tuiIconButton
-        icon="taiga-logo"
-        type="button">
+        icon="taiga-logo">
       </a>
       <div class="accesibility-help">
         <button
