+++ javascript/apps/taiga/src/assets/i18n/en-US.json
@@ -9,6 +9,7 @@
     "team": "Team",
     "cancel": "Cancel",
     "members": "Members",
+    "people": "People",
     "upgrade": "Upgrade",
     "owner": "Owner",
     "workspace": "Workspace",
@@ -83,6 +84,7 @@
   "common_project": {
     "create_first_project": "Create your first project",
     "create_new_project": "Add a project",
+    "create_project": "Create project",
     "project_actions": "Project actions",
     "card": {
       "accept_invite": "Accept invitation",
